Moved getCoherentRoutePathLinksList into RoutePathUtils
